+++ qtbook_cd.cc
@@ -494,8 +494,8 @@ void qtbook_cd::slotGo(void)
 	query.bindValue(19, oid);
       else if(qmain->getDB().driverName() == "QSQLITE")
 	{
-	  int value = misc_functions::getSqliteUniqueId(qmain->getDB(),
-							errorstr);
+	  qint64 value = misc_functions::getSqliteUniqueId(qmain->getDB(),
+							   errorstr);
 
 	  if(errorstr.isEmpty())
 	    query.bindValue(19, value);
