+++ qtbook_videogame.cc
@@ -455,8 +455,8 @@ void qtbook_videogame::slotGo(void)
 	query.bindValue(18, oid);
       else if(qmain->getDB().driverName() == "QSQLITE")
 	{
-	  int value = misc_functions::getSqliteUniqueId(qmain->getDB(),
-							errorstr);
+	  qint64 value = misc_functions::getSqliteUniqueId(qmain->getDB(),
+							   errorstr);
 
 	  if(errorstr.isEmpty())
 	    query.bindValue(18, value);
